% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.contigs.R
\name{add_contig_ctypes}
\alias{add_contig_ctypes}
\title{add_contig_ctypes}
\usage{
add_contig_ctypes(
  calns,
  refseq.fn = system.file("extdata", "reference.sequences.rds", package = "loosends"),
  remove.decoy = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{calns}{(data.table of contig alignments)}

\item{refseq.fn}{(path to key of reference sequences)}

\item{remove.decoy}{(logical) remove alignments to decoy sequences? default FALSE}

\item{verbose}{(logical) default FALSE}
}
\value{
data.table with new columns c_type and c_spec
}
\description{
Given data.table of contig alignments, add c_type and c_spec labels based on reference sequence
}
