% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{assemble_loose_reads}
\alias{assemble_loose_reads}
\title{assemble_loose_reads

this function loads reads surrounding a loose end and their mates,
assembles reads into contigs, aligns contigs to references,
and parses alignments to categorize the loose end}
\usage{
assemble_loose_reads(
  li,
  loose.reads.dt,
  ref_dir = system.file("extdata", "hg19_looseends", package = "loosends"),
  ref_obj = NULL,
  verbose = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{loose.reads.dt}{data.table output from loose.reads}

\item{verbose}{optional, default=FALSE}

\item{loose.end.gr}{GRanges}
}
\value{
data.table with assembled contigs
}
\description{
assemble_loose_reads

this function loads reads surrounding a loose end and their mates,
assembles reads into contigs, aligns contigs to references,
and parses alignments to categorize the loose end
}
