% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie_utils.R
\name{bowtie_aln}
\alias{bowtie_aln}
\title{bowtie_aln}
\usage{
bowtie_aln(
  reads,
  ref.seq = NULL,
  ref.dir = NULL,
  ref.basename = NULL,
  outdir = "./",
  verbose = FALSE
)
}
\arguments{
\item{reads}{(data.table with columns qname, seq)}

\item{ref.seq}{(data.table with columns qname, seq)}

\item{ref.dir}{(character) path to directory with reference fastas}

\item{ref.basename}{(character) basename of reference fasta}

\item{outdir}{(character) path to temporary output directory}

\item{verbose}{(logical)  default FALSE}
}
\value{
GRanges with alignment
}
\description{
Perform unpaired (single end) realignment of reads to either a supplied sequence or reference FASTA
}
