% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie_utils.R
\name{bowtie_index}
\alias{bowtie_index}
\title{bowtie_index}
\usage{
bowtie_index(fasta.fn = NA_character_, ref.id = "target", verbose = FALSE)
}
\arguments{
\item{fasta.fn}{(character) path to fasta}

\item{ref.id}{(character) basename of reference, default "target"}

\item{verbose}{(logical) default FALSE}
}
\value{
directory containing indexed FASTA
}
\description{
bowtie_index
}
