% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contig.assembly.R
\name{build_contigs}
\alias{build_contigs}
\title{build_contigs}
\usage{
build_contigs(
  reads.dt,
  col = "seed.frame",
  max.iter = 3,
  max.unaln = 3,
  verbose = FALSE
)
}
\arguments{
\item{reads.dt}{(must have column seed.frame)}

\item{col}{(column name containing sequences for alignment) e.g. seed.frame}

\item{max.iter}{(numeric) max number of iterations for assembly}

\item{max.unaln}{(numeric) max number of unaligned reads after assembly}

\item{verbose}{(logical) default FALSE}
}
\description{
Assemble reads surrounding seed window supplied in seed.frame column of data table
Performs alignment for a max number of iterations or until all reads are aligned
}
