% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.contigs.R
\name{build_contigs_wrapper}
\alias{build_contigs_wrapper}
\title{build_contigs_wrapper}
\usage{
build_contigs_wrapper(
  gr,
  reads.dt,
  ref,
  window = 5000,
  assembly.region = 1000,
  stride = 500,
  pseudo.contigs = TRUE,
  low.mappability.gr = GRanges(),
  unassembled.gr = GRanges(),
  verbose = FALSE
)
}
\arguments{
\item{gr}{(GRanges) which breakpoint do we care about}

\item{reads.dt}{(data.table)}

\item{ref}{(BWA object)}

\item{window}{(numeric) how many bases around window to consider for assembly (default 5e3)}

\item{assembly.region}{(numeric) how many bases at a time to consider for assembly (default 1e3)}

\item{stride}{(numeric) stride for tiling window (default 500)}

\item{pseudo.contigs}{(logical) include pseudo-contigs from discordant read clustering? default TRUE}

\item{low.mappability.gr}{(GRanges) path to GRanges containing low mappability bases}

\item{unassembled.gr}{(GRanges) path to GRanges containing unassembled regions}

\item{verbose}{(logical) default FALSE}
}
\value{
data.table with contig alignments
}
\description{
build_contigs_wrapper
}
