% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.contigs.R
\name{build_pseudo_contigs}
\alias{build_pseudo_contigs}
\title{build_pseudo_contigs}
\usage{
build_pseudo_contigs(
  reads.dt,
  col = "seed.frame",
  qcol = "qual",
  scol = "seed",
  ccol = "concord",
  min.mapq = 30,
  pad = 1000,
  min.count.thresh = 5,
  verbose = FALSE
)
}
\arguments{
\item{reads.dt}{(must have column seed frame or $col containing sequence and column seed or $scol containing whether or not the read is a seed read)}

\item{col}{(character) field with character of seed reading frame, default "seed.frame"}

\item{qcol}{(character) field indicating base quality, default "qual"}

\item{scol}{(character) field indicating whether read is seed, default "seed"}

\item{ccol}{(character) field indicating whether read is concordant, default "concord"}

\item{min.mapq}{(numeric) minimum mapq (default 1)}

\item{pad}{(numeric) pad for reducing alignments, default 1e3}

\item{min.count.thresh}{(numeric) minimum number of reads aligning to the pseudo-contig (default 3)}

\item{verbose}{(logical) default FALSE}
}
\value{
character vector of candidate contig sequences
}
\description{
Check whether there are three (default) read pairs where the mate aligns within 1 kbp of each other

If so returns the coordinates of the MATE (mate only)
}
