% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contig.assembly.R
\name{call_loose_end2}
\alias{call_loose_end2}
\title{call_loose_end2}
\usage{
call_loose_end2(
  li,
  ri,
  id,
  concat.bwa,
  human.bwa,
  window = 5000,
  low.mappability.gr.fn = system.file("extdata", "wide.multimapping.ranges.rds",
    package = "loosends"),
  unassembled.gr.fn = system.file("extdata", "assembly.gaps.ranges.rds", package =
    "loosends"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{li}{(data.table, GRanges, or character coercible to GRanges)}

\item{ri}{(reads from loosereads_wrapper)}

\item{id}{(character) sample id}

\item{concat.bwa}{(RSeqLib::BWA) reference BWA object with foreign sequences}

\item{human.bwa}{(RSeqLib::BWA) reference BWA with human only sequences}

\item{window}{(numeric) window in BP for building contigs, default 5e3}

\item{low.mappability.gr.fn}{(character) path to low mappability ranges}

\item{unassembled.gr.fn}{(character) path to low mappability ranges}

\item{verbose}{(logical) default FALSE}

\item{...}{(additional inputs to various helper functions)}
}
\value{
list with four items
\itemize{
\item call
\item contigs (curated contigs)
\item distal.bps (distal breakends)
\item all.bps (all breakends)
\item all.tigs (all contigs)
}
}
\description{
labels a single loose end
}
