% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{call_loose_end_wrapper}
\alias{call_loose_end_wrapper}
\title{call_loose_end_wrapper}
\usage{
call_loose_end_wrapper(
  id = "",
  le.dt = data.table(),
  reads.dt = data.table(),
  concat.bwa = NULL,
  human.bwa = NULL,
  concat.fn = NA_character_,
  pad = 5000,
  mix.tn = TRUE,
  minimap = FALSE,
  outdir = "./",
  max.reads = 5000,
  verbose = FALSE
)
}
\arguments{
\item{le.dt}{(from prep_loose_ends)}

\item{reads.dt}{(from loosereads)}

\item{concat.bwa}{(BWA from RSeqLib) concatenated fasta}

\item{human.bwa}{(BWA from RSeqLib)}

\item{pad}{(window for local assembly)}

\item{mix.tn}{(logical) mix tumor and normal reads? default TRUE}

\item{minimap}{(logical) default FALSE}

\item{max.reads}{(numeric) max reads per loose end before downsampling (default 5000)}

\item{verbose}{}
}
\value{
list with three items:
$call table with calls
$filtered.contigs data.table with filtered contigs
$all.contigs data.table with all contigs
}
\description{
call loose end given reads
}
