% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{caller}
\alias{caller}
\title{caller}
\usage{
caller(
  li,
  calns = NULL,
  insert = 750,
  pad = NULL,
  uannot = NULL,
  ref_dir = system.file("extdata", "hg19_looseends", package = "loosends")
)
}
\arguments{
\item{li}{data.table loose end to evaluate}

\item{calns}{optional, data.table of contig alignments, default=NULL (will not parse alignments)}

\item{insert}{optional, integer pad representing insert length in bp to identify seed alignments, default=750}

\item{pad}{optional, window around loose end to allow contig seed windows, default=1e3}

\item{uannot}{optional, GRanges of unmappable annotations, default bin/101.unmappable.annotations.rds}

\item{ref_dir}{optional, path to directory of unzipped reference tarball, default assumes 'package/extdata/hg19_looseends'}
}
\description{
parses contig alignments to assign loose end to category and describe repeat types
}
