% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contig.assembly.R
\name{check_contig_concordance}
\alias{check_contig_concordance}
\title{check_contig_concordance}
\usage{
check_contig_concordance(calns, verbose = FALSE)
}
\arguments{
\item{calns}{(data.table of contig alignments) needs to have columns "tumor.specific" and "tumor.support"}

\item{verbose}{(logical) default FALSE}
}
\value{
data.table with one row
with the following columns
\itemize{
\item any.alt (logical)
\item somatic.alt (logical)
\item junction (logical) are any of the contigs high-mapq junctions?
\item complex (logical) are any of the contigs high-mapq phased complex rearrangements (no junctions)?
\item single breakend (logical) no high-mapq junctions or phased complex rearrangements, but there are ALT contigs.
\item insertion (logical) do the junctions contain insertions? (NA if not a junction)
\item homology (logical) do the junctions contain nonzero breakend homology? (NA if not a junction)
\item telomeric (logical) do any tumor-specific contigs contain telomeric sequence?
\item jstring (character) grl of the consensus junction (NA if not a junctions)
}
}
\description{
check_contig_concordance
}
