% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.contigs.R
\name{check_contigs_for_telomeres}
\alias{check_contigs_for_telomeres}
\title{check_contigs_for_telomeres}
\usage{
check_contigs_for_telomeres(calns, verbose = FALSE)
}
\arguments{
\item{calns}{(data.table from contig alignments)}

\item{verbose}{(logical) default FALSE}
}
\value{
data.table with columns:
\itemize{
\item query_c_telomere
\item query_g_telomere
\item aln_c_telomere
\item aln_g_telomere
\item c_telomere (any alignment for the qname contains a C telomere)
\item g_telomere (any alignment for the qname contains a G telomere)
}
}
\description{
check contigs for telomeres
}
