% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contig.support.R
\name{check_distal_only_contig_support}
\alias{check_distal_only_contig_support}
\title{check_distal_only_contig_support}
\usage{
check_distal_only_contig_support(
  calns,
  reads.dt,
  ref.pad = 5000,
  seed.pad = 0,
  verbose = FALSE
)
}
\arguments{
\item{calns}{(data.table) represents contig alignment for a SINGLE contig}

\item{reads.dt}{(data.table) reads corresponding to the loose end associated with this contig}

\item{ref.pad}{(numeric) pad (bp) around seed region to get reference alignment, default 5 kbp}

\item{seed.pad}{(numeric) number of base pairs around peak to search for supporting reads, default 0}

\item{verbose}{(logical) default FALSE}
}
\description{
Gets supporting reads for contigs with distal-only alignments
This is likely to cause false positives for split contigs so use only with distal-only contigs
}
