% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosereads.R
\name{check_file}
\alias{check_file}
\title{check_file}
\usage{
check_file(fn = NULL)
}
\arguments{
\item{fn}{}
}
\value{
logical, TRUE if file exists and is nonempty
}
\description{
Check if file supplied is nonempty
}
