% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{contig.support}
\alias{contig.support}
\title{contig.support}
\usage{
contig.support(
  reads,
  contig,
  ref = NULL,
  min.bases = 20,
  min.aligned.frac = 0.95
)
}
\arguments{
\item{reads}{GRanges of seed reads and mates}

\item{contig}{data.table of reference alignments for a single contig}

\item{ref}{optional, character of reference sequence surrounding loose end, default=NULL}

\item{min.bases}{optional, minimum aligned bases from a read to count for support, default=20}

\item{min.aligned.frac}{optional, minimum fraction of aligned bases from a read to count for support, default=0.95}
}
\description{
quantifies preferential read support for contig over reference
}
