% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contig.assembly.R
\name{contig_is_single_chunk}
\alias{contig_is_single_chunk}
\title{contig_is_single_chunk}
\usage{
contig_is_single_chunk(calns)
}
\arguments{
\item{calns}{(GRanges) contig to reference alignment}
}
\description{
return TRUE if contig is chimeric
}
