% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{.realign}
\alias{.realign}
\title{.realign}
\usage{
.realign(reads, ref, gg = NULL, filter = TRUE)
}
\arguments{
\item{reads}{GRanges or data.table of reads from BAM file with $seq in original reading frame}

\item{ref}{reference genome for realignment}

\item{gg}{optional, gGraph of sample used to identify sequences fitted in graph, default=NULL}

\item{filter}{optional, filter=TRUE will return loose read pairs only, filter=FALSE returns all, default=TRUE}
}
\description{
realigns reads and their mates single-end to attach individual MAPQs
}
