% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{.sample.spec}
\alias{.sample.spec}
\title{.sample.spec}
\usage{
.sample.spec(le, bam, pad)
}
\arguments{
\item{le}{GRanges or data.table of loose ends}

\item{bam}{path to BAM file}

\item{pad}{integer width of padding to add around loose ends}
}
\description{
loads reads and mates for a single sample (tumor or normal)
}
