% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{.waviness}
\alias{.waviness}
\title{.waviness}
\usage{
.waviness(
  x,
  y,
  min.thresh = 5000,
  max.thresh = 1e+05,
  spar = 0.5,
  smooth = TRUE,
  filter = rep(FALSE, length(x)),
  trim = 10
)
}
\description{
quantifies autocorrelation in coverage
}
