% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{eighteenmer}
\alias{eighteenmer}
\title{eighteenmer}
\usage{
eighteenmer(gorc = c("g", "c", "both"))
}
\arguments{
\item{gorc}{'g' returns 18mers of G-heavy strand only 'c' returns 18mers of C-heavy strand only 'both' returns 18mers of G-heavy strand and 18mers of C-heavy strand, default='both'}
}
\description{
generate a PDict of telomeric 18mers
}
