% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterloose.R
\name{filter_loose}
\alias{filter_loose}
\title{filter_loose}
\usage{
filter_loose(
  jabba_rds,
  coverage = "/dev/null",
  norm.field = "reads.corrected",
  mask = "~/projects/gGnome/files/zc_stash/tile.mask.rds",
  mask_pad = 0,
  norm_thresh = 0.6,
  min_size = 0,
  subset = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{jabba_rds}{(character) path to JaBbA}

\item{coverage}{(character) path to coverage file with gc-adjusted counts}

\item{norm.field}{(character) normal count field (reads.corrected)}

\item{mask}{(character) path to coverage mask}

\item{mask_pad}{(numeric) pad on distance to mask (in practice just for annotation purposes)}

\item{norm_thresh}{(numeric) 0.6 (from filter.loose)}

\item{min_size}{(numeric) minimum size in BP}

\item{subset}{(logical) subset loose ends to include only valid ones before returning}

\item{verbose}{(logical)}
}
\description{
filter_loose
}
