% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contig.assembly.R
\name{find_telomeres}
\alias{find_telomeres}
\title{find_telomeres}
\usage{
find_telomeres(seq = character(), gorc = "g", verbose = FALSE)
}
\arguments{
\item{seq}{(character) vector of sequences}

\item{gorc}{(either g, c, or both, input to eighteenmer)}

\item{verbose}{(logical) default FALSE}
}
\value{
logical with length of seq, indicating whether that sequence contains a c or g telomere
}
\description{
find_telomeres
}
