% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contig.assembly.R
\name{grab_contig_breakends}
\alias{grab_contig_breakends}
\title{grab_contig_breakends}
\usage{
grab_contig_breakends(
  calns,
  seed,
  reduce.pad = 20,
  seed.pad = 1000,
  verbose = FALSE
)
}
\arguments{
\item{calns}{(GRanges) contig alignment}

\item{seed}{(GRanges) seed region of the contig}

\item{reduce.pad}{(numeric) reduce multipmapping regions using this pad (default 20)}

\item{seed.pad}{(numeric) bp pad to be considered lying outside peak (default 1e3)}

\item{verbose}{(logical)}
}
\value{
data.table with seqnames, start, end, strand, and seed.side (indicating breakpoint)
strand is in breakend orientation (so if there were a junction it would be consistent with the junction)
}
\description{
Get seed side and mate side breakends for a single contig.

Assumes that a unique qname is supplied.
Also assumes that the alignment is split.

If more than one contig is supplied, throws an exception.
If the alignment is not split, returns an empty data table with a warning.
}
