% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contig.assembly.R
\name{grab_contig_breakends_wrapper}
\alias{grab_contig_breakends_wrapper}
\title{grab_contig_breakends_wrapper}
\usage{
grab_contig_breakends_wrapper(calns, name.field = "name", verbose = FALSE)
}
\arguments{
\item{calns}{(data.table) contig alignments}

\item{name.field}{(character) default name}

\item{verbose}{(logical)}
}
\description{
grab_contig_breakends_wrapper
}
