% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contig.assembly.R
\name{grab_distal_breakends}
\alias{grab_distal_breakends}
\title{grab_distal_breakends}
\usage{
grab_distal_breakends(
  bps,
  mapq.thresh = 60,
  specificity.thresh = 0.05,
  pad = 1000
)
}
\arguments{
\item{bps}{(data.table) data.table representing breakpoints coercible to GRanges. bps needs to have columns mapq, tumor.count, and normal.count representing supporting barcodes for that contig.}

\item{mapq.thresh}{(default 60)}

\item{specificity.thresh}{(numeric) cutoff for tumor-specific contigs}

\item{pad}{(numeric) default 1 kbp}
}
\value{
a data.table with filtered breakends and additional metadata columns
}
\description{
Grab candidate distal breakends based on tumor-specificity

If there are tumor-specific breakends:
If there are MAPQ = 60 breakends all within 1kbp: returns a single breakend (the first one)
If there are MAPQ = 60 breakends but not within 1 kbp: return MAPQ = 60 breakends
Otherwise, return all breakends
}
