% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosereads.R
\name{grab_looseread_params}
\alias{grab_looseread_params}
\title{grab_looseread_params}
\usage{
grab_looseread_params(
  gr = GRanges(),
  bam = character(),
  pad = 5000,
  mate_pad = 150,
  verbose = FALSE
)
}
\arguments{
\item{gr}{(GRanges) stranded GRanges representing loose ends}

\item{bam}{(character) path to bam file}

\item{pad}{(numeric) pad around loose ends}

\item{mate_pad}{(numeric) pad around mate windows}

\item{verbose}{(logical)}
}
\value{
list with names:
\itemize{
\item $ranges (GRanges of windows)
\item $qnames (character vector of loose read qnames)
}
}
\description{
Grab ranges and qnames of reads near loose ends + their mates
(Including ranges for split reads)
}
