% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosereads.R
\name{grab_loosereads}
\alias{grab_loosereads}
\title{grab_loosereads}
\usage{
grab_loosereads(
  bam = NA_character_,
  ranges = GRanges(),
  qnames = character(),
  outdir = "./",
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{bam}{(character) path to BAM file}

\item{ranges}{(GRanges)}

\item{qnames}{(character)}

\item{outdir}{(character) path to dump results}

\item{overwrite}{(logical) overwrite existing files? default TRUE}

\item{verbose}{(logical) default FALSE}
}
\value{
path to indexed BAM file with all required reads
}
\description{
get the reads + mates associated with loose ends
}
