% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{grab_ref_obj}
\alias{grab_ref_obj}
\title{grab_ref_obj}
\usage{
grab_ref_obj(ref.dir = NA_character_)
}
\arguments{
\item{ref.dir}{(character) path to reference directory}
}
\value{
list with names rep, human, polyA, microbe
}
\description{
grab_ref_obj
}
