% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contig.assembly.R
\name{grab_seed_frame}
\alias{grab_seed_frame}
\title{grab_seed_frame}
\usage{
grab_seed_frame(
  reads.dt,
  seed.gr,
  seq.field = "reading.frame",
  max.n = 0,
  verbose = FALSE
)
}
\arguments{
\item{reads.dt}{(data.table) must have columns seqnames, start, end, strand, reading.frame, qname}

\item{seed.gr}{(GRanges) (stranded) seed window for assembly}

\item{seq.field}{(character) where is the seq found? default reading.frame}

\item{max.n}{(numeric) maximum number of low-quality "N" bases (default 0)}

\item{verbose}{(logical) default FALSE}
}
\description{
Get reads and their mates overlapping the seed window
Designate one member of the pair the "seed" read and the other pair the "mate" read
Reverse complement the "mate" read so that the reference frame is the same
}
