% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contig.assembly.R
\name{label_contigs}
\alias{label_contigs}
\title{label_contigs}
\usage{
label_contigs(
  le,
  bps,
  mapq.gr,
  unmappable.gr,
  mappability.thresh = 0.5,
  specificity.thresh = 0
)
}
\arguments{
\item{bps}{(data.table) data.table representing breakpoints coercible to GRanges. bps needs to have columns mapq, tumor.count, and normal.count representing supporting barcodes for that contig.}

\item{mapq.gr}{(GRanges) tiled granges with field mapq0.frac}

\item{unmappable.gr}{(GRanges) ranges that are considered CN unmappable}

\item{mappability.thresh}{(numeric) 0.1 for mappable categorization}

\item{specificity.thresh}{(numeric) cutoff for tumor-specific contigs}

\item{li}{(data.table) data.table coercible to GRanges representing loose end}

\item{returns}{a list with two items:
\itemize{
\item call (data.table with a single row and metadata column call)
\item breakpoints (data.table labeled breakpoints. empty table if no contigs)
}}
}
\description{
Labels the category of each contig associated with a loose end, as well as the overall
category of the loose end. (e.g. T0/T1/T2/MYS)
}
