% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{loose.reads}
\alias{loose.reads}
\title{loose.reads}
\usage{
loose.reads(
  le,
  tbam,
  pad = 25000,
  nbam = NULL,
  ref = system.file("extdata", "hg19_looseends", "human_g1k_v37_decoy.fasta", package =
    "loosends"),
  filter = TRUE,
  gg = NULL,
  verbose = FALSE
)
}
\arguments{
\item{le}{GRanges or data.table windows around which to sesarch for reads}

\item{tbam}{character path to tumor BAM file}

\item{pad}{optional, integer padding around le windows, default=25e3}

\item{nbam}{optional, character path to normal BAM file, default=NULL}

\item{ref}{optional, BWA object of reference genome for realignment or path to fasta, default=package/extdata/hg19_looseends/human_g1k_v37_decoy.fasta}

\item{filter}{optional, logical filter=TRUE returns loose read pairs only, filter=FALSE returns all read pairs annotated with logical $loose column, default=T}

\item{gg}{optional, gGraph corresponding to sample, used to identify sequences fitted in graph, default=NULL}

\item{verbose}{optional, default=FALSE}
}
\description{
loads and reads and their mates from given windows and realigns single end for read-specific MAPQ
}
