% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosereads.R
\name{loose.reads2}
\alias{loose.reads2}
\title{loose.reads2}
\usage{
loose.reads2(
  tbam,
  taln,
  nbam = NA,
  naln = NA,
  id = "",
  ref = system.file("extdata", "hg19_looseends", "human_g1k_v37_decoy.fasta", package =
    "loosends"),
  filter = TRUE,
  gg = NULL,
  verbose = FALSE
)
}
\arguments{
\item{tbam}{character path to tumor BAM file}

\item{nbam}{optional, character path to normal BAM file, default=NULL}

\item{ref}{optional, BWA object of reference genome for realignment or path to fasta, default=package/extdata/hg19_looseends/human_g1k_v37_decoy.fasta}

\item{filter}{optional, logical filter=TRUE returns loose read pairs only, filter=FALSE returns all read pairs annotated with logical $loose column, default=T}

\item{gg}{optional, gGraph corresponding to sample, used to identify sequences fitted in graph, default=NULL}

\item{verbose}{optional, default=FALSE}
}
\value{
data.table of reads relaigned to the specified reference
}
\description{
Performs the following:
\itemize{
\item loads reads from a specified window around loose end
\item identifies any split reads and loads the "sides" of those reads as well
\item loads the mates of reads around the loose end (by qname)
\item realigns all of these reads
}

The supplied BAM files are expected to have tag SA (indicated split alignments)
The functionality is identical to loose.reads; however, the BAMs are expected to be prefiltered to contain only the relevant QNAMEs via samtools to avoid slow filtering in R for reads with many mate windows
}
