% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{match.seq}
\alias{match.seq}
\title{match.seq}
\usage{
match.seq(query, subject)
}
\arguments{
\item{query}{PDict of query sequences to match}

\item{subject}{DNAStringSet of sequences to parse for matches}
}
\description{
returns logical vector of length subject indicating whether matches were found with any query
}
