% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{merge.telomeres}
\alias{merge.telomeres}
\title{merge.telomeres}
\usage{
\method{merge}{telomeres}(treps, x, ch, out.dt, c_spec = "telomere")
}
\description{
construct a GRanges in contig coordinates for telomere motif matches
}
