% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{minimap_index}
\alias{minimap_index}
\title{minimap_index}
\usage{
minimap_index(fasta = NA_character_, outdir = "./", verbose = FALSE)
}
\arguments{
\item{fasta}{(character) path to fasta to pre-index}

\item{outdir}{(character) path to target directory (created if not already exists)}

\item{verbose}{(logical) default FALSE}
}
\value{
path to minimap indexed reference (.mmi)
}
\description{
minimap_index
}
