% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{munch}
\alias{munch}
\title{munch}
\usage{
munch(reads, query = NULL)
}
\arguments{
\item{reads}{GRanges or data.table containing $seq field with character sequence}

\item{query}{optional, PDict of motifs, default= 18mers of telomeric motifs (both strands)}
}
\description{
returns logical vector of length reads
indicating does read contain any match to query
searches for matches on same strand as input seq
}
