% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosereads.R
\name{parse_realignment}
\alias{parse_realignment}
\title{parse_realignment}
\usage{
parse_realignment(reads, aln_bam, filter = FALSE, verbose = FALSE)
}
\arguments{
\item{reads}{(GRanges) reads from .sample_spec}

\item{aln_bam}{(character) bam realigned to reference}

\item{filter}{(logical) default FALSE}

\item{verbose}{(logical) default FALSE}
}
\value{
data.table of realigned reads
}
\description{
takes two BAMs, one with all reads and mates, and one with realignments
and merges them, marking reads which did not realign uniquely to the reference
}
