% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{prep_loose_ends}
\alias{prep_loose_ends}
\title{prep_loose_ends}
\usage{
prep_loose_ends(li, id = "")
}
\arguments{
\item{li}{(GRanges, character, or data.table)}

\item{id}{(character) sample ID}
}
\value{
data.table of loose ends with columns sample and leix
}
\description{
prep_loose_ends
}
