% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{process.loose.ends}
\alias{process.loose.ends}
\title{process.loose.ends}
\usage{
process.loose.ends(
  le,
  tbam,
  nbam = NULL,
  id = NULL,
  outdir = NULL,
  mc.cores = 1,
  ref_dir = system.file("extdata", "hg19_looseends", package = "loosends"),
  verbose = FALSE
)
}
\arguments{
\item{le}{data.table or GRanges of loose ends to analyze}

\item{tbam}{character path to tumor BAM file}

\item{nbam}{optional, character path to normal BAM file, default=NULL}

\item{id}{optional, character sample id, default=NULL}

\item{outdir}{optional, character path to output directory, default=NULL (will not write output files if outdir=NULL)}

\item{mc.cores}{optional, parallel cores for building contigs per loose end, default=1}

\item{ref_dir}{optional, path to directory of unzipped reference tarball, default assumes 'package/extdata/hg19_looseends'}

\item{verbose}{optional, default=FALSE}
}
\description{
this function categorizes an input set of loose ends
runs process.single.end for each input loose end
}
