% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{process_loose_ends}
\alias{process_loose_ends}
\title{process_loose_ends}
\usage{
process_loose_ends(
  id = "",
  ranges = GRanges(),
  tbam = "/dev/null",
  nbam = "/dev/null",
  ref_dir = "/dev/null",
  outdir = "./",
  read_pad = 5000,
  assembly_pad = 1000,
  verbose = FALSE
)
}
\arguments{
\item{id}{(character) sample ID}

\item{ranges}{(GRanges)}

\item{tbam}{(character) path to tumor bam}

\item{nbam}{(character) path to normal bam}

\item{ref_dir}{(character) path to reference directory}

\item{outdir}{(character) directory for storing outputs}

\item{read_pad}{(numeric) default 5000}

\item{assembly_pad}{(numeric) default 1000}

\item{verbose}{(logical) default TRUE}
}
\value{
list with two items
\itemize{
\item call
\item contigs
\item support
}
}
\description{
process_loose_ends
}
