% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.contigs.R
\name{qc_contigs}
\alias{qc_contigs}
\title{qc_contigs}
\usage{
qc_contigs(
  calns,
  seed.gr,
  low.mappability.gr = GRanges(),
  unassembled.gr = GRanges(),
  athresh = 20,
  seed.pad = 1000
)
}
\arguments{
\item{calns}{(data.table) alignment of possibly multiple contigs}

\item{seed.gr}{(GRanges) genomic ranges of seed region used to assembled contigs}

\item{low.mappability.gr}{(GRanges) ranges of low mappability bases}

\item{unassembled.gr}{(GRanges) ranges of unassembled bases}

\item{athresh}{(numeric) # matching bases needed for alignment}

\item{seed.pad}{(numeric) need at least this many bps from seed region to be considered a "distal" aln}
}
\value{
data.table.
}
\description{
Given (possibly multiple) contigs and their alignments
Identifies contigs corresponding with ALT alleles and removes
}
