% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.contigs.R
\name{qc_single_contig}
\alias{qc_single_contig}
\title{qc_single_contig}
\usage{
qc_single_contig(
  calns.dt,
  seed.gr,
  low.mappability.gr = GRanges(),
  unassembled.gr = GRanges(),
  athresh = 20,
  seed.pad = 1000,
  mapq.thresh = 60
)
}
\arguments{
\item{seed.gr}{(GRanges) genomic range of seed region for contig assembly}

\item{low.mappability.gr}{(GRanges) GRanges of low mappability bases}

\item{unassembled.gr}{(GRanges) GRanges of unassembled bases}

\item{athresh}{(numeric) minimum number of aligned bases a valid alignment (default 20 bp)}

\item{seed.pad}{(numeric) a distal alignment must be at least this many bases away from the seed. should be close to maximum plausible insert size (default 1000 bp)}

\item{calns}{(data.table) containing a single contig and its alignments. This must be unique.}

\item{maqp.thresh}{(numeric) default 60, minimum alignment MAPQ to be considered a 'high MAPQ' alignment}
}
\value{
data.table with the following added columns
\itemize{
\item outside.seed (logical)
\item outside.stranded.seed (logical)
\item alength (numeric) : total length of alignment
\item single.chunk (logical)
\item keep (logical) - this indicates whether the contig should be kept for further analysis
\item fbi (logical) - is the contig a foldback-inversion
\item unmapped bases (logical) does the contig have > athresh unmapped bases?
\item junction (logical) does the contig represent a junction?
\item complex (logical) does the contig represent a phased complex rearrangement?
\item homology (numeric) number of base pairs of breakend homology (only applicable to junctions)
\item insertion (numeric) number of inserted base pairs between breakends (only applicable to junctions)
\item high.mapq (logical)
}
}
\description{
Takes a single contig (one qname only) and the seed locus from which the contig was assembled.
Annotates the sequences to which the contig was aligned (e.g. human/viral/etc.)
Identify whether the contig represents an ALT allele

Potential ALT alleles include:
\itemize{
\item contigs not aligning to the seed region at all
\item contigs aligning to the seed region but with a substantial number of unmapped bases
\item contigs representing junctions, including INV, TRA, DEL, DEUP
\item contigs representing phased complex rearrangements
}
}
