% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contigsupport.R
\name{read_support}
\alias{read_support}
\title{read_support}
\usage{
read_support(
  le.dt = data.table(),
  reads.dt = data.table(),
  contigs.dt = data.table(),
  calns = data.table(),
  fasta = "/dev/null",
  ref.bwa = NULL,
  chimeric = TRUE,
  strict = TRUE,
  min.bases = 20,
  min.aligned.frac = 0.95,
  isize.diff = 1000,
  seed.pad = 150,
  all.reads = FALSE,
  bowtie = FALSE,
  outdir = "./",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{le.dt}{(data.table) (single row output from process_loose_ends)}

\item{reads.dt}{(data.table) data.table of reads coercible to GRanges}

\item{contigs.dt}{(data.table) data.table with colnames track, peak, qname. all qnames should be identical.}

\item{calns}{(data.table) contig alignments, coercible to GRanges}

\item{seed.pad}{(numeric) pad around seed for getting supporting reads}

\item{all.reads}{(logical) return all input reads or just the chunks supporing the contig? (default FALSE)}

\item{bowtie}{(logical) use bowtie to align reads? if TRUE, fasta must be provided}

\item{outdir}{(character) output directory for temporary files}

\item{verbose}{(logical)}

\item{...}{additional parameters to contig.support}
}
\value{
data.table of candidate reads
if all.reads is FALSE and there are no supporting reads an empty table is returned
if all.reads is TRUE the table will have column $supporting indicating whether read suupports given contig
}
\description{
wrapper around contig.support for easy(-ish) debugging

like contig.support, the input contigs must be nonempty and contain a single qname
}
