% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosereads.R
\name{realign_loosereads}
\alias{realign_loosereads}
\title{realign_loosereads}
\usage{
realign_loosereads(
  bam,
  ref = system.file("extdata", "hg19_looseends", "human_g1k_v37_decoy.fasta", package =
    "loosends"),
  outdir = "./",
  verbose = FALSE
)
}
\arguments{
\item{bam}{(character) path to bam file}

\item{ref}{(character) path to BWA indexed reference}

\item{outdir}{(character) output directory to dump results}

\item{verbose}{(logical)}

\item{overwrite}{(logical) overwrite existing analysis}
}
\value{
path to output bam file with realigned reads
}
\description{
Realign reads to selected reference
}
