% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie_utils.R
\name{seqs_to_fasta}
\alias{seqs_to_fasta}
\title{seqs_to_fasta}
\usage{
seqs_to_fasta(seqs.dt, fasta.fn, format = "fasta", verbose = FALSE)
}
\arguments{
\item{seqs.dt}{(data.table) must have columns qname and fastq}

\item{fasta.fn}{(character) fasta file name to write to}

\item{format}{(fasta or fastq)}

\item{verbose}{(logical) default FALSE}
}
\value{
fasta.fn
}
\description{
seqs_to_fasta
}
