% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{transform}
\alias{transform}
\title{transform}
\usage{
transform(seq, s, e)
}
\arguments{
\item{seq}{character seqnames}

\item{s}{integer start}

\item{e}{integer end}
}
\description{
this function allows conversion between data.table and gr.sum
(used for convenience with a data.table by= argument)
}
