% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{update.call}
\alias{update.call}
\title{update.call}
\usage{
update.call(dt)
}
\description{
modifies call string if logical column values have changed
(used when supplementing contig calls with discordant read pairs)
}
