% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contig.assembly.R
\name{align_contigs}
\alias{align_contigs}
\title{align_contigs}
\usage{
align_contigs(
  tigs,
  ref,
  refseq.fn = system.file("extdata", "reference.sequences.rds", package = "loosends"),
  primary.only = TRUE,
  remove.decoy = TRUE,
  keep.unaligned = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{tigs}{(character)}

\item{ref}{(RSeqLib::BWA)}

\item{refseq.fn}{(character) path to seqnames annotation file (included)}

\item{primary.only}{(logical) default TRUE}

\item{remove.decoy}{(logical) exclude alignment to decoy sequences? default TRUE}

\item{keep.unaligned}{(logical) keep unaligend contigs?? default FALSE}

\item{verbose}{(logical) default FALSE}
}
\value{
data.table with contig alignment and metadata fields c_type and c_spec
}
\description{
align_contigs
}
