% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.contigs.R
\name{align_contigs}
\alias{align_contigs}
\title{align_contigs}
\usage{
align_contigs(
  tigs,
  ref,
  primary.only = TRUE,
  remove.decoy = TRUE,
  keep.unaligned = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{tigs}{(character)}

\item{ref}{(RSeqLib::BWA)}

\item{primary.only}{(logical) default TRUE}

\item{keep.unaligned}{(logical) keep unaligend contigs?? default FALSE}

\item{verbose}{(logical) default FALSE}

\item{refseq.fn}{(character) path to seqnames annotation file (included)}
}
\value{
data.table with contig alignment and metadata fields seq, query.seq
}
\description{
Align contigs (character vector) to reference
Then annotate the segments of the alignment according to reference seqnames

Relies on seqnames --> annotation mapping included as part of the package...
}
