% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{call_loose_end}
\alias{call_loose_end}
\title{call_loose_end}
\usage{
call_loose_end(
  li,
  ri,
  ref_obj = NULL,
  pad = 1000,
  mix.tn = TRUE,
  max.reads = 2500,
  verbose = FALSE
)
}
\arguments{
\item{li}{(data.table) data.table coercible to GRanges, must have metatdata $leix and $sample}

\item{ri}{(data.table) loose reads table prepped (e.g. from prep_loose_reads)}

\item{ref_obj}{(character) or a list of BWA indices, if ref_dir not provided}

\item{pad}{(numeric) window size for local assembly, default 1 kbp}

\item{mix.tn}{(logical) mix tumor/normal reads before assembly}

\item{max.reads}{(logical) maximum number of reads before downsampling (default 5000)}

\item{verbose}{}
}
\value{
list with entries
\itemize{
\item all.contigs (data.table)
\item wide.contigs (data.table or NULL)
\item call (data.table)
}
}
\description{
This function calls a loose end given reads from near to the loose end
}
