% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{call_loose_end_wrapper}
\alias{call_loose_end_wrapper}
\title{call_loose_end_wrapper}
\usage{
call_loose_end_wrapper(
  id = "",
  le.dt = data.table(),
  reads.dt = data.table(),
  ref_obj = NULL,
  pad = 5000,
  mix.tn = TRUE,
  max.reads = 5000,
  verbose = FALSE
)
}
\arguments{
\item{le.dt}{(from prep_loose_ends)}

\item{reads.dt}{(from loosereads)}

\item{ref_obj}{(reference object)}

\item{pad}{(window for local assembly)}

\item{mix.tn}{(logical)}

\item{max.reads}{(numeric) max reads per loose end before downsampling (default 5000)}

\item{verbose}{}
}
\value{
list with three items:
$call table with calls
$filtered.contigs data.table with filtered contigs
$all.contigs data.table with all contigs
}
\description{
call loose end given reads
}
