% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contig.support.R
\name{check_contig_support_wrapper}
\alias{check_contig_support_wrapper}
\title{check_contig_support_wrapper}
\usage{
check_contig_support_wrapper(
  calns,
  reads.dt,
  ref,
  name.field = "name",
  min.supp.reads = 2,
  max.normal.reads = 5,
  min.read.ratio = 5,
  verbose = FALSE
)
}
\arguments{
\item{calns}{(data.table) alignment for POSSIBLY MULTIPLE contigs}

\item{reads.dt}{(data.table)}

\item{ref}{(BWA)}

\item{name.field}{(character) default "name"}

\item{min.supp.reads}{(numeric) min number of supporting reads to be considered as having read support (default 2)}

\item{max.normal.reads}{(numeric) maximum accepted number of supporting reads in normal (default 5)}

\item{min.read.ratio}{(numeric) minimum ratio of tumor to normal reads (default 5)}

\item{verbose}{(logical) default FALSE}
}
\description{
Given a set of filtered contigs, checks support for each contig using the appropriate function
e.g. if the contig is chimeric, uses check_split_contig_support
whereas, if the contig contains the distal side only, uses check_distal_only_contig_support

Produces a data table with the name of each contig and the number of tumor/normal supporting qnames
}
