% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{.realign}
\alias{.realign}
\title{.realign}
\usage{
.realign(reads, ref, gg = NULL, filter = TRUE, verbose = F, chunksize = 5e+05)
}
\arguments{
\item{reads}{(GRanges or data.table) reads from BAM file with $seq in original reading frame}

\item{ref}{(BWA object from RSeqLib) reference genome for realignment}

\item{gg}{(gGraph) sample used to identify sequences fitted in graph, default=NULL}

\item{filter}{(logical) return loose reads pairs only? default TRUE}

\item{verbose}{optional, default=FALSE}

\item{chunksize}{(numeric) perform realignment in chunks to help with memory. default 1e6}
}
\description{
realigns reads and their mates single-end to attach individual MAPQs
}
