% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{filter.graph}
\alias{filter.graph}
\title{filter.graph}
\usage{
\method{filter}{graph}(gg, cov.rds, field = "ratio", verbose = F)
}
\arguments{
\item{gg}{gGraph of JaBbA genome graph (contains loose ends)}

\item{cov.rds}{character path to binned coverage data}

\item{field}{optional, default = "ratio", column in binned coverage data}
}
\value{
data.table containing a row for every loose end in gGraph gg with a logical column \code{true.pos} indicating whether each loose end has passed all filters (TRUE) or not (FALSE)
}
\description{
identifies quality loose ends from gGraph
}
