% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{filter.graph}
\alias{filter.graph}
\title{filter.graph}
\usage{
\method{filter}{graph}(
  gg,
  cov.rds,
  purity = NULL,
  ploidy = NULL,
  field = "ratio",
  PTHRESH = 3.4e-07,
  verbose = F
)
}
\arguments{
\item{gg}{gGraph of JaBbA genome graph (contains loose ends)}

\item{cov.rds}{character path to binned coverage data}

\item{purity}{optional, fractional purity of sample, default assumes 1}

\item{ploidy}{optional, ploidy of sample, default infers from gg}

\item{field}{optional, default = "ratio", column in binned coverage data}

\item{PTHRESH}{optional, threshold for GLM p-value for calling true positive loose ends, default=3.4e-7 provides consanguinity with large dataset bonferroni correction}

\item{verbose}{optional, default=FALSE}
}
\value{
data.table containing a row for every loose end in gGraph gg with a logical column \code{true.pos} indicating whether each loose end has passed all filters (TRUE) or not (FALSE)
}
\description{
identifies quality loose ends from gGraph
}
