% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{filter.loose}
\alias{filter.loose}
\title{filter.loose}
\usage{
\method{filter}{loose}(
  gg,
  cov.rds,
  l,
  purity = NULL,
  ploidy = NULL,
  field = "ratio",
  PTHRESH = 3.4e-07,
  verbose = F
)
}
\arguments{
\item{gg}{gGraph of JaBbA model}

\item{cov.rds}{character path to binned coverage data}

\item{l}{data.table of loose ends to evaluate}

\item{purity}{optional, fractional purity of sample, default assumes 1}

\item{ploidy}{optional, ploidy of sample, default infers from gg}

\item{field}{optional, column name in cov.rds, default="ratio"}

\item{PTHRESH}{optional, threshold for GLM p-value for calling true positive loose ends, default=3.4e-7 provides consanguinity with large dataset bonferroni correction}

\item{verbose}{optional, default=FALSE}
}
\value{
data.table containing a row for every input loose end and logical column \code{true.pos} indicating whether each loose end has passed all filters (TRUE) or not (FALSE)
}
\description{
analyze coverage surrounding given loose ends to evaluate quality
}
