% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{ggraph.loose.ends}
\alias{ggraph.loose.ends}
\title{ggraph.loose.ends}
\usage{
ggraph.loose.ends(
  gg,
  cov.rds,
  tbam,
  nbam = NULL,
  id = NULL,
  outdir = NULL,
  purity = NULL,
  ploidy = NULL,
  field = "ratio",
  PTHRESH = 3.4e-07,
  verbose = F,
  mc.cores = 1,
  ref_dir = system.file("extdata", "hg19_looseends", package = "loosends"),
  ref_obj = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{gg}{a gGraph built from a JaBbA model (ie with loose ends)}

\item{cov.rds}{character path to binned coverage file (should contain GRanges)}

\item{tbam}{character path to tumor BAM file}

\item{nbam}{optional, character path to normal BAM file, default=NULL}

\item{id}{optional, character sample id, default=NULL}

\item{outdir}{optional, character path to output directory, default=NULL (will not write output files if outdir=NULL)}

\item{purity}{optional, fractional purity of sample, default assumes 1}

\item{ploidy}{optional, ploidy of sample, default infers from gg}

\item{field}{optional, name of informative column in cov.rds, default="ratio"}

\item{PTHRESH}{optional, threshold for GLM p-value for calling true positive loose ends, default=3.4e-7 provides consanguinity with large dataset bonferroni correction}

\item{verbose}{optional, will print status messages, default=FALSE}

\item{mc.cores}{optional, parallel cores for building contigs per loose end, default=1}

\item{ref_dir}{optional, path to directory of unzipped reference tarball, default assumes 'package/extdata/hg19_looseends'}

\item{overwrite, }{optional, logical indicating whether to generate new output files if corresponding files have already been written to outdir, default=FALSE (will load existing files)}
}
\description{
this function processes a single gGraph to identify
high quality loose ends and categorize them
}
