% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosereads.R
\name{loose.reads2}
\alias{loose.reads2}
\title{loose.reads2}
\usage{
loose.reads2(
  tbam,
  taln,
  nbam = NA,
  naln = NA,
  id = "",
  filter = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{tbam}{(character) path to tumor BAM file}

\item{taln}{(character) path realigned tumor bam}

\item{nbam}{(character) optional, path to normal BAM file, default=NULL}

\item{naln}{(character) optional, realigned normal bam}

\item{id}{(character)}

\item{filter}{(logical) return loose read pairs only? default FALSE}

\item{verbose}{optional, default=FALSE}
}
\value{
data.table of reads relaigned to the specified reference
}
\description{
Performs the following:
\itemize{
\item loads reads from a specified window around loose end
\item identifies any split reads and loads the "sides" of those reads as well
\item loads the mates of reads around the loose end (by qname)
\item realigns all of these reads
}

The supplied BAM files are expected to have tag SA (indicated split alignments)
The functionality is identical to loose.reads; however, the BAMs are expected to be prefiltered to contain only the relevant QNAMEs via samtools to avoid slow filtering in R for reads with many mate windows
}
