% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosereads.R
\name{loosereads_wrapper}
\alias{loosereads_wrapper}
\title{loosereads_wrapper}
\usage{
loosereads_wrapper(
  ranges = GRanges(),
  tbam = "/dev/null",
  nbam = "/dev/null",
  ref = "/dev/null",
  bowtie = FALSE,
  bowtie.ref = "/dev/null",
  bowtie.dir = "/dev/null",
  id = "",
  outdir = "./",
  pad = 5000,
  mask = "/dev/null",
  cleanup = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{ranges}{(GRanges) loose end GRanges}

\item{tbam}{(character) path to normal bam}

\item{nbam}{(character) optional, path to normal bam}

\item{ref}{(character) path to reference .fasta}

\item{bowtie}{(logical) use bowtie for realignment?}

\item{bowtie.ref}{(character) basename of bowtie references}

\item{bowtie.dir}{(character) dirname of bowtie}

\item{id}{(character) sample name}

\item{outdir}{(character)}

\item{pad}{(numeric)}

\item{mask}{(character) path to GRanges mask}

\item{cleanup}{(logical) remove temporary files? default TRUE}

\item{verbose}{(logical)}
}
\value{
data.table of reads + mates
}
\description{
wrapper for obtaining reads + mates around loose ends and realigning
}
