% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{prep_loose_reads}
\alias{prep_loose_reads}
\title{prep_loose_reads

Get loose reads data table ready for assembly}
\usage{
prep_loose_reads(li, loose.reads.dt)
}
\arguments{
\item{loose.reads.dt}{}
}
\value{
data.table
}
\description{
prep_loose_reads

Get loose reads data table ready for assembly
}
