% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contig.assembly.R
\name{qc_contigs}
\alias{qc_contigs}
\title{qc_contigs}
\usage{
qc_contigs(calns, seed.gr, athresh = 20, seed.pad = 1000)
}
\arguments{
\item{calns}{(GRanges) alignment of possibly multiple contigs}

\item{athresh}{(numeric) # matching bases needed for alignment}

\item{seed.pad}{(numeric) need at least this many bps from seed region to be considered a "distal" aln}
}
\description{
Given contigs and their alignments, remove contigs that are uninformative for loose end classification
}
