% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{read.based}
\alias{read.based}
\title{read.based}
\usage{
read.based(li, ri, pad = NULL)
}
\arguments{
\item{li}{data.table loose end to classify}

\item{ri}{data.table read alignments}

\item{pad}{optional, padding around loose end li to search for discordant reads, default=1e3}
}
\description{
classifying loose end based on discordant read alignments by constructing pseudo-contigs based on consensus alignment patterns
}
