% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{read_support}
\alias{read_support}
\title{read_support}
\usage{
read_support(
  loose.end = NA,
  loose.end.str = NA_character_,
  reads.dt = data.table(),
  contigs.dt = data.table(),
  ref.bwa = NULL,
  fasta = NA_character_,
  pad = 1000,
  refseq.pad = 10000,
  seed.pad = 0,
  chimeric = TRUE,
  strict = TRUE,
  min.bases = 20,
  min.aligned.frac = 0.95,
  isize.diff = 1000,
  all.reads = FALSE,
  ref.coordinates = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{loose.end}{(GRanges) single loose end}

\item{loose.end.str}{(character) loose end as a string (one of loose.end/loose.end.str should be supplied.)}

\item{reads.dt}{(data.table) data.table of reads coercible to GRanges}

\item{contigs.dt}{(data.table) data.table with colnames track, peak, qname. all qnames should be identical.}

\item{pad}{(numeric) pad loose ends for getting the refseq?}

\item{seed.pad}{(numeric) pad around}

\item{all.reads}{(logical) return all input reads or just the chunks supporing the contig? (default FALSE)}

\item{ref.coordinates}{(logical) want reads in ref (vs contig) coordinates? default TRUE}

\item{verbose}{(logical)}

\item{...}{additional parameters to contig.support}
}
\value{
data.table of candidate reads
if all.reads is FALSE and there are no supporting reads an empty table is returned
if all.reads is TRUE the table will have column $supporting indicating whether read suupports given contig
}
\description{
wrapper around contig.support for easy(-ish) debugging

like contig.support, the input contigs must be nonempty and contain a single qname
}
