% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosends.R
\name{return.dt}
\alias{return.dt}
\title{return.dt}
\usage{
return.dt(
  reference,
  complex,
  missedj,
  novel,
  mystery,
  insertion,
  refmap = NULL,
  novmap = NULL,
  rrep = NULL,
  irep = NULL,
  nrep = NULL
)
}
\description{
constructs output data.table from logical arguments
}
